scripts/requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
faker>=22.0.0
python-dotenv>=1.0.0
requests>=2.31.0
